+++ src/main/java/de/samply/teiler/TeilerConst.java
@@ -62,7 +62,6 @@ public class TeilerConst {
 
   public static final Integer FIRST_ANONYM_ID = 1;
   public static final String DEFAULT_ANONYM_ELEMENT = "VAL";
-  public static final String DEFAULT_QUERY_ID = "EMBEDDED_QUERY";
   public static final String CHILD_FHIR_PATH_HEAD = "/";
 
 }